:tada: Initial with state from Netz39 and Freifunk
We are using Ansible to administrate machines for the Freifunk Magdeburg
project and internal machines of the Netz39 Hackerpace since 2020.
Quite some Docker containers are running on that infrastructure.
Setting up the Docker runtime environment was implemented in a role in
each of the site repositories.  The implementation is almost identical
in both repos, and to be reusable we merge them in this standalone role.

Link: https://docs.ansible.com/
Link: https://github.com/FreifunkMD/backend-ffmd
Co-authored-by: David Kilias <david.kilias@gmail.com>
Co-authored-by: Stefan Haun <tux@netz39.de>