:page_facing_up: Make project reuse compliant
The main license of the project is MIT, which applies to tasks,
handlers, and the like.  Documentation (README.md) is licensed as CC-BY.
Configuration files of the project itself are "licensed" CC0.  The jinja
templates are also for configuration files, but for the target.  Jinja
comments are used here, especially for the json file because JSON does
not allow comments.  Files taken from the old Freifunk Magdeburg and
Netz39 site repos got the year of creation and original author as
copyright line.

Link: https://reuse.software/