Always install the prune service, but limit timer activation
Change the prune service installationso that the systemd service unit
is always available, but timer activation depends on the configuration.

This allows to trigger image pruning from sources other than the timer,
even if the timer is not used.

Behavior of the installation does not change, other than availability
of the prune systemd unit.