Add python3-standard-pipes to Debian 13 install
The [standard-pipes](https://pypi.org/project/standard-pipes/) library has been removed from the Python standard libraries in [PEP 594](https://peps.python.org/pep-0594/), but is needed by the docker tasks.

Add the Debian package in Debian 13 installations to satisfy the dependency.

The library is not needed to install Docker, but will be missing when subsequent tasks try to interact with Docker.